src/penroselamarck/db/python-requirements.txt
sqlalchemy>=2.0
alembic>=1.12,<2
asyncpg>=0.29,<1
pydantic==2.11.9
pydantic-settings>=2.1,<3
psycopg[binary]>=3.1,<4
openlineage-python>=1.9,<2
